+++ packages/at_secondary_server/test/stats_verb_test.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:math';
 
 import 'package:at_persistence_secondary_server/at_persistence_secondary_server.dart';
 import 'package:at_secondary/src/caching/cache_manager.dart';
@@ -472,6 +473,41 @@ void main() {
     });
 
     test(
+        'A test to validate commit entries when commit log entry count is greater than default sync buffer zie',
+        () async {
+      secondaryPersistenceStore = SecondaryPersistenceStoreFactory.getInstance()
+          .getSecondaryPersistenceStore(alice);
+      LastCommitIDMetricImpl.getInstance().atCommitLog =
+          secondaryPersistenceStore!.getSecondaryKeyStore()!.commitLog;
+      var lastCommitId =
+          await LastCommitIDMetricImpl.getInstance().getMetrics();
+      print('lastCommitId before op: $lastCommitId');
+      var randomString = Uuid().v4();
+      int phoneNumber = 1234;
+      int min = 5;
+      int max = 100;
+      // generate a random integer between 5 and 100
+      int randomNumber = min + Random().nextInt(max - min) + 1;
+      for (int i = 1; i <= randomNumber; i++) {
+        phoneNumber = phoneNumber + i;
+        await secondaryPersistenceStore!.getSecondaryKeyStore()!.put(
+            '@alice:phone-${randomString}_$i@alice',
+            AtData()..data = phoneNumber.toString());
+      }
+      lastCommitId = await LastCommitIDMetricImpl.getInstance().getMetrics();
+      var latestCommitIdForEachKey =
+          await LatestCommitEntryOfEachKey().getMetrics();
+      Map<String, dynamic> latestCommitIdMap =
+          jsonDecode(latestCommitIdForEachKey);
+      for (int i = 1; i <= randomNumber; i++) {
+        expect(
+            latestCommitIdMap
+                .containsKey('@alice:phone-${randomString}_$i@alice'),
+            true);
+      }
+    });
+
+    test(
         'A test to verify latest commitId among enrolled namespaces is returned',
         () async {
       await secondaryPersistenceStore!
@@ -555,6 +591,18 @@ void main() {
       var lastCommitId =
           await LastCommitIDMetricImpl.getInstance().getMetrics(regex: 'buzz');
       expect(lastCommitId, '2');
+    });
+    test('A test to check LatestCommitEntryOfEachKey for empty commit log',
+        () async {
+      secondaryPersistenceStore = SecondaryPersistenceStoreFactory.getInstance()
+          .getSecondaryPersistenceStore(alice);
+      LastCommitIDMetricImpl.getInstance().atCommitLog =
+          secondaryPersistenceStore!.getSecondaryKeyStore()!.commitLog;
+      var latestCommitIdForEachKey =
+          await LatestCommitEntryOfEachKey().getMetrics();
+      Map<String, dynamic> latestCommitIdMap =
+          jsonDecode(latestCommitIdForEachKey);
+      expect(latestCommitIdMap.isEmpty, true);
     });
     tearDown(() async => await verbTestsTearDown());
   });